Add missing unsaved files

+++ src/linker/Linker.Steps/DescriptorMarker.cs
@@ -72,7 +72,7 @@ namespace Mono.Linker.Steps
 				}
 
 				if (!foundMatch) {
-					LogWarning (new DiagnosticString (DiagnosticId.XmlCouldNotFindAnyTypeInNamespace).GetMessage (fullname), (int) DiagnosticId.XmlCouldNotFindAnyTypeInNamespace, namespaceNav);
+					LogWarning (new DiagnosticString (DiagnosticId.XmlCouldNotFindAnyTypeInNamespace).GetMessage (fullname), DiagnosticId.XmlCouldNotFindAnyTypeInNamespace, namespaceNav);
 				}
 			}
 		}
@@ -102,11 +102,11 @@ namespace Mono.Linker.Steps
 			TypePreserve preserve = GetTypePreserve (nav);
 			switch (preserve) {
 			case TypePreserve.Fields when !type.HasFields:
-				LogWarning (new DiagnosticString (DiagnosticId.TypeHasNoFieldsToPreserve).GetMessage (type.GetDisplayName ()), (int) DiagnosticId.TypeHasNoFieldsToPreserve, nav);
+				LogWarning (new DiagnosticString (DiagnosticId.TypeHasNoFieldsToPreserve).GetMessage (type.GetDisplayName ()), DiagnosticId.TypeHasNoFieldsToPreserve, nav);
 				break;
 
 			case TypePreserve.Methods when !type.HasMethods:
-				LogWarning (new DiagnosticString (DiagnosticId.TypeHasNoMethodsToPreserve).GetMessage (type.GetDisplayName ()), (int) DiagnosticId.TypeHasNoMethodsToPreserve, nav);
+				LogWarning (new DiagnosticString (DiagnosticId.TypeHasNoMethodsToPreserve).GetMessage (type.GetDisplayName ()), DiagnosticId.TypeHasNoMethodsToPreserve, nav);
 				break;
 
 			case TypePreserve.Fields:
@@ -148,7 +148,7 @@ namespace Mono.Linker.Steps
 		protected override void ProcessField (TypeDefinition type, FieldDefinition field, XPathNavigator nav)
 		{
 			if (_context.Annotations.IsMarked (field))
-				LogWarning (new DiagnosticString (DiagnosticId.XmlDuplicatePreserveMember).GetMessage (field.FullName), (int) DiagnosticId.XmlDuplicatePreserveMember, nav);
+				LogWarning (new DiagnosticString (DiagnosticId.XmlDuplicatePreserveMember).GetMessage (field.FullName), DiagnosticId.XmlDuplicatePreserveMember, nav);
 
 			_context.Annotations.Mark (field, new DependencyInfo (DependencyKind.XmlDescriptor, _xmlDocumentLocation), GetMessageOriginForPosition (nav));
 		}
@@ -156,7 +156,7 @@ namespace Mono.Linker.Steps
 		protected override void ProcessMethod (TypeDefinition type, MethodDefinition method, XPathNavigator nav, object? customData)
 		{
 			if (_context.Annotations.IsMarked (method))
-				LogWarning (new DiagnosticString (DiagnosticId.XmlDuplicatePreserveMember).GetMessage (method.GetDisplayName ()), (int) DiagnosticId.XmlDuplicatePreserveMember, nav);
+				LogWarning (new DiagnosticString (DiagnosticId.XmlDuplicatePreserveMember).GetMessage (method.GetDisplayName ()), DiagnosticId.XmlDuplicatePreserveMember, nav);
 
 			_context.Annotations.MarkIndirectlyCalledMethod (method);
 			_context.Annotations.SetAction (method, MethodAction.Parse);
@@ -213,7 +213,7 @@ namespace Mono.Linker.Steps
 		protected override void ProcessEvent (TypeDefinition type, EventDefinition @event, XPathNavigator nav, object? customData)
 		{
 			if (_context.Annotations.IsMarked (@event))
-				LogWarning (new DiagnosticString (DiagnosticId.XmlDuplicatePreserveMember).GetMessage (@event.FullName), (int) DiagnosticId.XmlDuplicatePreserveMember, nav);
+				LogWarning (new DiagnosticString (DiagnosticId.XmlDuplicatePreserveMember).GetMessage (@event.FullName), DiagnosticId.XmlDuplicatePreserveMember, nav);
 
 			ProcessMethod (type, @event.AddMethod, nav, customData);
 			ProcessMethod (type, @event.RemoveMethod, nav, customData);
@@ -225,7 +225,7 @@ namespace Mono.Linker.Steps
 			string[] accessors = fromSignature ? GetAccessors (nav) : _accessorsAll;
 
 			if (_context.Annotations.IsMarked (property))
-				LogWarning (new DiagnosticString (DiagnosticId.XmlDuplicatePreserveMember).GetMessage (property.FullName), (int) DiagnosticId.XmlDuplicatePreserveMember, nav);
+				LogWarning (new DiagnosticString (DiagnosticId.XmlDuplicatePreserveMember).GetMessage (property.FullName), DiagnosticId.XmlDuplicatePreserveMember, nav);
 
 			if (Array.IndexOf (accessors, "all") >= 0) {
 				ProcessMethodIfNotNull (type, property.GetMethod, nav, customData);
@@ -236,12 +236,12 @@ namespace Mono.Linker.Steps
 			if (property.GetMethod != null && Array.IndexOf (accessors, "get") >= 0)
 				ProcessMethod (type, property.GetMethod, nav, customData);
 			else if (property.GetMethod == null)
-				LogWarning (new DiagnosticString (DiagnosticId.XmlCouldNotFindGetAccesorOfPropertyOnType).GetMessage (property.Name, type.FullName), (int) DiagnosticId.XmlCouldNotFindGetAccesorOfPropertyOnType, nav);
+				LogWarning (new DiagnosticString (DiagnosticId.XmlCouldNotFindGetAccesorOfPropertyOnType).GetMessage (property.Name, type.FullName), DiagnosticId.XmlCouldNotFindGetAccesorOfPropertyOnType, nav);
 
 			if (property.SetMethod != null && Array.IndexOf (accessors, "set") >= 0)
 				ProcessMethod (type, property.SetMethod, nav, customData);
 			else if (property.SetMethod == null)
-				LogWarning (new DiagnosticString (DiagnosticId.XmlCouldNotFindSetAccesorOfPropertyOnType).GetMessage (property.Name, type.FullName), (int) DiagnosticId.XmlCouldNotFindSetAccesorOfPropertyOnType, nav);
+				LogWarning (new DiagnosticString (DiagnosticId.XmlCouldNotFindSetAccesorOfPropertyOnType).GetMessage (property.Name, type.FullName), DiagnosticId.XmlCouldNotFindSetAccesorOfPropertyOnType, nav);
 		}
 
 		static bool IsRequired (XPathNavigator nav)
